# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the Ticket Place App",
        "contact": {
            "email": "ticket.place@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://apid.c4m.mg/ticket-place-app/public",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Get All Tags List",
                "description": "return list of the all event tag",
                "operationId": "4d0f4baf9e6fa992c0d85fbf140ebf17",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "action unauthorized"
                    },
                    "403": {
                        "description": "action forbiden"
                    },
                    "404": {
                        "description": "aucun résultat trouvé"
                    },
                    "500": {
                        "description": "erreur serveur"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Create new Tag",
                "description": "Create new Tag",
                "operationId": "a0a246f29dcb119a2189fd230fc10028",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "concert"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "donné incomplet"
                    },
                    "401": {
                        "description": "action unauthorized"
                    },
                    "403": {
                        "description": "action forbiden"
                    },
                    "500": {
                        "description": "erreur serveur"
                    }
                }
            }
        },
        "/api/tags/{tag_id}": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Get One Tag Info",
                "description": "return info of the tag",
                "operationId": "17cf4f4a882aae3d887866484660df56",
                "parameters": [
                    {
                        "name": "tag_id",
                        "in": "path",
                        "description": "ID of the tag to show",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "action unauthorized"
                    },
                    "403": {
                        "description": "action forbiden"
                    },
                    "404": {
                        "description": "aucun résultat trouvé"
                    },
                    "500": {
                        "description": "erreur serveur"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tags"
                ],
                "summary": "Update Tag",
                "description": "Update Tag label",
                "operationId": "3a64462cc7467aedc95b80bf19a299de",
                "parameters": [
                    {
                        "name": "tag_id",
                        "in": "path",
                        "description": "ID of the tag to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "bal masqué"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "donné incomplet"
                    },
                    "401": {
                        "description": "action unauthorized"
                    },
                    "403": {
                        "description": "action forbiden"
                    },
                    "500": {
                        "description": "erreur serveur"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tags"
                ],
                "summary": "Delete tag",
                "description": "Delete tag",
                "operationId": "d24105e20ffcde076b873eaec69e2965",
                "parameters": [
                    {
                        "name": "tag_id",
                        "in": "path",
                        "description": "ID of the tag to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "action not authorized"
                    },
                    "403": {
                        "description": "action forbiden"
                    },
                    "500": {
                        "description": "erreur serveur"
                    }
                }
            }
        },
        "/api/my-tickets": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user's tickets",
                "description": "return the list of user's tickets",
                "operationId": "7cb72bb0c912d3f203b9c01811317ff9",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "aucun résultat trouvé"
                    },
                    "500": {
                        "description": "erreur serveur"
                    }
                }
            }
        },
        "/api/events/{event_id}/tickets": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get Event's Tickets",
                "description": "return list of Event's tickets",
                "operationId": "2538e567322eb94714ad69baffae7ee9",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "ID of the Event",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "action unauthorized"
                    },
                    "404": {
                        "description": "aucun résultat trouvé"
                    },
                    "500": {
                        "description": "erreur serveur"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}